+++ wasm/browser.js
@@ -1,0 +1,1 @@
+export * from 'imgkit-wasm32-wasi'